add ajax to message table

+++ Feasablty study/Domin/Entites/Competitors.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string CompetitorsName { get; set; }
         public string Products { get; set; }
-        public decimal CompetitorsPrice { get; set; }
+        public double CompetitorsPrice { get; set; }
         public int AmounrAdded { get; set; }
         public int MarketStudyId { get; set; }
         public Market_study market_Study { get; set; }
